Create Tea Store Sales and Orders.
First I created a store data base with the different types of tea sold, the price per box, the number of items sold and the amount of pending orders to sell. 

Then I created tables separating the teas by price and order the tables by the most to the least sold on each price point.

To analyze the amount of sales I created a table that multiplied the retail price and the amount of boxes sold and ordered them by ascending price, having the least sold on the top.

I also created somewhat of a queue with the total amount of pending orders.